requirements.txt
Flask==2.3.3
pandas==2.2.3
python-dotenv==1.0.0
Jinja2==3.1.2
requests==2.32.3
